Separated assignment_source_filter into the app and the task, extracted stages_from_config_reader
